apps/backend/requirements.txt
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Flask==2.2.5
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
SQLAlchemy==2.0.37
typing_extensions==4.12.2
Werkzeug==3.1.3
Flask-JWT-Extended==4.4.4
Flask-Migrate==3.1.0
Flask-Mail==0.10.0
Python-Dotenv==1.0.1
Flask-Bcrypt==1.0.1
groq==0.11.0
requests==2.31.0